Add background worker for the read models
At the moment the Cashier service only cares about new products so it
can use them when placing orders.
The worker will subscribe to the $ce-product EventStore Stream and use
ProductCreated events to hydrate the Product read model repository.
We're not using any checkpoint, so it will read all events since beginning
of times, meaning it will try to add all items to the repository again.
It shouldn't be a problem for now, since we'll use an in-memory repository.